Workaround for the system weird state caused by RootPort Memory Access
The problem with current RootPort Memory Access tests (448 and 449)
may be that it doesn't use a proper way to change the memory window
(memory limit). At least, it doesn't consider the end-point device
behind the root port. That caused system  to enter a weird state
to cause fake failures with some other test cases (e.g. Timer Wakeup
tests) or cause some devices non-functional.

The workaround is to directly access the address out of the memory
window without changing the size the memory window.
Note this workaround doesn't completely work. It may still accidentally
access an inaccessible region to still cause the same problem (We may
get the value like 0xDEADBEEF from the address out of memory window).
To improve this workaround, we have to find a way to know where is the
inaccessible regions. In my experiment, with this workaround, the issue
with test 448 got fixed, but it still fails with test 449 as it accesses
an inaccessible region.

Signed-off-by: Sunny Wang <sunny.wang@arm.com>